# setDate refactoring to fix HHVM failing tests