--- a/benchmarks/Chronos/SetDateBench.php
+++ b/benchmarks/Chronos/SetDateBench.php
@@ -1,0 +1,16 @@
+<?php
+namespace Cake\Chronos\Benchmarks\Chronos;
+
+use Cake\Chronos\Chronos;
+
+class SetDateBench
+{
+    /**
+     * @Revs(100000)
+     * @Iterations(5)
+     */
+    public function benchSetDate()
+    {
+       $date = (new Chronos())->setDate(1, 1, 1);
+    }
+}