--- a/benchmarks/Chronos/SetDateBench.php
+++ b/benchmarks/Chronos/SetDateBench.php
@@ -8,9 +8,10 @@ class SetDateBench
     /**
      * @Revs(100000)
      * @Iterations(5)
+     * @return void
      */
     public function benchSetDate()
     {
-       $date = (new Chronos())->setDate(1, 1, 1);
+        $date = (new Chronos())->setDate(1, 1, 1);
     }
 }