--- a/benchmarks/DateTime/SetDateBench.php
+++ b/benchmarks/DateTime/SetDateBench.php
@@ -1,0 +1,16 @@
+<?php
+namespace Cake\Chronos\Benchmarks\MutableDateTime;
+
+use Cake\Chronos\MutableDateTime;
+
+class SetDateBench
+{
+    /**
+     * @Revs(100000)
+     * @Iterations(5)
+     */
+    public function benchSetDate()
+    {
+       $date = (new MutableDateTime())->setDate(1, 1, 1);
+    }
+}