--- a/benchmarks/DateTime/SetDateBench.php
+++ b/benchmarks/DateTime/SetDateBench.php
@@ -8,9 +8,10 @@ class SetDateBench
     /**
      * @Revs(100000)
      * @Iterations(5)
+     * @return void
      */
     public function benchSetDate()
     {
-       $date = (new MutableDateTime())->setDate(1, 1, 1);
+        $date = (new MutableDateTime())->setDate(1, 1, 1);
     }
 }