--- a/phpbench.json
+++ b/phpbench.json
@@ -1,0 +1,3 @@
+{
+    "bootstrap": "vendor/autoload.php"
+}