--- a/phpcs.xml.dist
+++ b/phpcs.xml.dist
@@ -6,6 +6,7 @@
 
  <file>./src</file>
  <file>./tests</file>
+ <file>./benchmarks</file>
 
  <arg phpcs-only="true" value="p"/>
  <arg name="extensions" value="php"/>