--- a/src/Traits/ModifierTrait.php
+++ b/src/Traits/ModifierTrait.php
@@ -113,17 +113,20 @@ trait ModifierTrait
      */
     public function setDate($year, $month, $day)
     {
-        // Workaround for PHP issue.
-        $date = $this->modify('+0 day');
+        return $this->modify('+0 day')->setDateParent($year, $month, $day);
+    }
 
-        if ($this instanceof DateTimeImmutable) {
-            // Reflection is necessary to access the parent method
-            // of the immutable object
-            $method = new \ReflectionMethod('DateTimeImmutable', 'setDate');
-
-            return $method->invoke($date, $year, $month, $day);
-        }
-
+    /**
+     * Just calling to parent setDate
+     * It used in overwritten setDate
+     *
+     * @param int $year The year to set.
+     * @param int $month The month to set.
+     * @param int $day The day to set.
+     * @return static
+     */
+    private function setDateParent($year, $month, $day)
+    {
         return parent::setDate($year, $month, $day);
     }