@Ovsyanka 2017-04-05
Actualy i realised that HHVM tests looks different after rebasing on master (i think because my #141 pull-request). But i had a conversation with the @markstory and he tald me that HHVM actually not planned to be supported.
Considering that, the pull-request now becomes just about changing the workaround implementation.
I think it looks a slightly more clear now, but it is more "opinion matter" now.