@lorenzo 2017-04-06
I agree that the implementation is more clear now, I liked the change regardless of it being targeted for HHVM